// frontend/src/services/utils/interfaces/employee.interface.tsx
export interface EmployeeInterface {
    cedula: string;
    nombres: string;
    apellidos: string;
    correo: string;
    fechaNacimiento: string;
    direccion: string;
    movil: string;
    estaVacunado: boolean,
    nombreUsuario: string;
    password: string;
    rol: string;
    tipoVacuna: string;
    fechaVacunacion: string;
    numeroDosis: number;
}